+++ internal/provider/provider.go
@@ -120,6 +120,7 @@ func (p *GigahostProvider) Resources(_ context.Context) []func() resource.Resour
 		NewSSHKeyResource,
 		NewServerResource,
 		NewServerRoutedLayer3IPv4Resource,
+		NewServerIPReverseResource,
 	}
 }